:art: Refactor Post package to follow Clean Architecture
