+++ src/main/java/com/trip/IronBird_Server/post/adapter/mapper/PostMapper.java
@@ -1,0 +1,22 @@
+package com.trip.IronBird_Server.post.adapter.mapper;
+
+import com.trip.IronBird_Server.post.adapter.dto.PostDto;
+import com.trip.IronBird_Server.post.domain.Post;
+import org.springframework.stereotype.Component;
+
+@Component
+public class PostMapper {
+
+    public PostDto postDto(Post post){
+        return PostDto.builder()
+                .id(post.getId())
+                .title(post.getTitle())
+                .detail(post.getDetail())
+                .userName(post.getUser().getName())
+                .planId(post.getPlan().getId())
+                .createTime(post.getCreateTime())
+                .modifyTime(post.getModifyTime())
+                .build();
+    }
+
+}
